Extract SinglePlayer and its related variables & fields into a separate class.
